fix: output in console message
